server/requirements.txt
email_validator==2.1.0
Flask==3.1.0
Flask_Cors==4.0.0
Flask_Limiter==3.5.0
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.1.0
python_bcrypt==0.3.2
